requirements.txt
Pillow >= 9.1.0
PyYAML >= 6.0
Shapely >= 1.7.1
lxml >= 4.8.0
matplotlib >= 3.6.2
opencv_python_headless >= 4.5.4.60
openslide_python
pandas >= 1.4.2
rectangle_packer == 2.0.2
scikit_image >= 0.19.2
scipy >= 1.8.0
tabulate >= 0.8.9
tqdm >= 4.64.0
